refactor: refactor factory pattern
