+++ internal/store/factory.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package factory
+package store
 
 import (
 	"fmt"
@@ -21,8 +21,8 @@ import (
 	"github.com/notaryproject/ratify-go"
 )
 
-// NewStoreOptions defines the options for creating a new store.
-type NewStoreOptions struct {
+// NewOptions defines the options for creating a new [ratify.Store].
+type NewOptions struct {
 	// Type represents a specific implementation of a store. Required.
 	Type string `json:"type"`
 
@@ -34,10 +34,10 @@ type NewStoreOptions struct {
 }
 
 // registeredStores saves the registered store factories.
-var registeredStores map[string]func(*NewStoreOptions) (ratify.Store, error)
+var registeredStores map[string]func(*NewOptions) (ratify.Store, error)
 
 // RegisterStore registers a store factory to the system.
-func RegisterStoreFactory(storeType string, create func(*NewStoreOptions) (ratify.Store, error)) {
+func RegisterStoreFactory(storeType string, create func(*NewOptions) (ratify.Store, error)) {
 	if storeType == "" {
 		panic("store type cannot be empty")
 	}
@@ -45,7 +45,7 @@ func RegisterStoreFactory(storeType string, create func(*NewStoreOptions) (ratif
 		panic("store factory cannot be nil")
 	}
 	if registeredStores == nil {
-		registeredStores = make(map[string]func(*NewStoreOptions) (ratify.Store, error))
+		registeredStores = make(map[string]func(*NewOptions) (ratify.Store, error))
 	}
 	if _, registered := registeredStores[storeType]; registered {
 		panic(fmt.Sprintf("store factory type %s already registered", storeType))
@@ -53,8 +53,35 @@ func RegisterStoreFactory(storeType string, create func(*NewStoreOptions) (ratif
 	registeredStores[storeType] = create
 }
 
-// NewStore creates a new Store instance based on the provided options.
-func NewStore(opts *NewStoreOptions) (ratify.Store, error) {
+// New creates a new [ratify.StoreMux] instance where each store is registered
+// for its respective scopes.
+func New(opts []*NewOptions, globalScopes []string) (ratify.Store, error) {
+	if len(opts) == 0 {
+		return nil, fmt.Errorf("no store options provided")
+	}
+	storeMux := ratify.NewStoreMux()
+	for _, storeOptions := range opts {
+		if len(storeOptions.Scopes) == 0 {
+			// if no scopes are provided, use the global scopes of the executor.
+			storeOptions.Scopes = globalScopes
+		}
+		store, err := newStore(storeOptions)
+		if err != nil {
+			return nil, fmt.Errorf("failed to create store for type %q: %w", storeOptions.Type, err)
+		}
+		for _, scope := range storeOptions.Scopes {
+			if err = storeMux.Register(scope, store); err != nil {
+				return nil, fmt.Errorf("failed to register store for scope %q: %w", scope, err)
+			}
+		}
+	}
+
+	return storeMux, nil
+}
+
+// newStore creates a new [ratify.Store] instance based on the provided options
+// and will be used to register the store in the [ratify.StoreMux].
+func newStore(opts *NewOptions) (ratify.Store, error) {
 	if opts.Type == "" {
 		return nil, fmt.Errorf("store type is not provided in the store options")
 	}
