+++ internal/store/filesystemocistore/register.go
@@ -21,14 +21,14 @@ import (
 	"os"
 
 	"github.com/notaryproject/ratify-go"
-	"github.com/notaryproject/ratify/v2/internal/store/factory"
+	"github.com/notaryproject/ratify/v2/internal/store"
 )
 
 const filesystemOCIStoreType = "filesystem-oci-store"
 
 func init() {
 	// Register the filesystem OCI store factory
-	factory.RegisterStoreFactory(filesystemOCIStoreType, func(opts *factory.NewStoreOptions) (ratify.Store, error) {
+	store.RegisterStoreFactory(filesystemOCIStoreType, func(opts *store.NewOptions) (ratify.Store, error) {
 		if opts.Parameters == nil {
 			return nil, fmt.Errorf("store parameters are required")
 		}
