+++ internal/store/registrystore/register_test.go
@@ -18,18 +18,19 @@ package registrystore
 import (
 	"testing"
 
-	"github.com/notaryproject/ratify/v2/internal/store/factory"
+	"github.com/notaryproject/ratify/v2/internal/store"
+	_ "github.com/notaryproject/ratify/v2/internal/store/credentialprovider/static" // Register the static credential provider factory
 )
 
 func TestNewStore(t *testing.T) {
 	tests := []struct {
 		name      string
-		opts      *factory.NewStoreOptions
+		opts      *store.NewOptions
 		expectErr bool
 	}{
 		{
 			name: "Unsupported params type",
-			opts: &factory.NewStoreOptions{
+			opts: &store.NewOptions{
 				Type:       registryStoreType,
 				Parameters: make(chan int),
 			},
@@ -37,7 +38,7 @@ func TestNewStore(t *testing.T) {
 		},
 		{
 			name: "Malformed JSON params",
-			opts: &factory.NewStoreOptions{
+			opts: &store.NewOptions{
 				Type:       registryStoreType,
 				Parameters: "{invalid json",
 			},
@@ -45,7 +46,7 @@ func TestNewStore(t *testing.T) {
 		},
 		{
 			name: "Missing credential provider",
-			opts: &factory.NewStoreOptions{
+			opts: &store.NewOptions{
 				Type: registryStoreType,
 				Parameters: map[string]interface{}{
 					"plainHttp": true,
@@ -55,7 +56,7 @@ func TestNewStore(t *testing.T) {
 		},
 		{
 			name: "Invalid credential provider type",
-			opts: &factory.NewStoreOptions{
+			opts: &store.NewOptions{
 				Type: registryStoreType,
 				Parameters: map[string]interface{}{
 					"credential": map[string]interface{}{
@@ -67,7 +68,7 @@ func TestNewStore(t *testing.T) {
 		},
 		{
 			name: "Valid registry params with static credential provider",
-			opts: &factory.NewStoreOptions{
+			opts: &store.NewOptions{
 				Type: registryStoreType,
 				Parameters: map[string]interface{}{
 					"plainHttp":        true,
@@ -85,7 +86,7 @@ func TestNewStore(t *testing.T) {
 		},
 		{
 			name: "Valid registry params with minimal config",
-			opts: &factory.NewStoreOptions{
+			opts: &store.NewOptions{
 				Type: registryStoreType,
 				Parameters: map[string]interface{}{
 					"credential": map[string]interface{}{
@@ -98,7 +99,7 @@ func TestNewStore(t *testing.T) {
 		},
 		{
 			name: "Empty credential provider options",
-			opts: &factory.NewStoreOptions{
+			opts: &store.NewOptions{
 				Type: registryStoreType,
 				Parameters: map[string]interface{}{
 					"credential": map[string]interface{}{},
@@ -110,7 +111,7 @@ func TestNewStore(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			store, err := factory.NewStore(test.opts)
+			store, err := store.New([]*store.NewOptions{test.opts}, nil)
 			if (err != nil) != test.expectErr {
 				t.Errorf("expected error: %v, got: %v", test.expectErr, err)
 			}
@@ -127,7 +128,7 @@ func TestNewStore(t *testing.T) {
 func TestRegistryStoreFactory(t *testing.T) {
 	// Test that the factory is properly registered
 	t.Run("Factory is registered", func(t *testing.T) {
-		opts := &factory.NewStoreOptions{
+		opts := &store.NewOptions{
 			Type: registryStoreType,
 			Parameters: map[string]interface{}{
 				"credential": map[string]interface{}{
@@ -137,7 +138,7 @@ func TestRegistryStoreFactory(t *testing.T) {
 			},
 		}
 
-		_, err := factory.NewStore(opts)
+		_, err := store.New([]*store.NewOptions{opts}, nil)
 		if err != nil {
 			t.Errorf("factory should be registered and functional, got error: %v", err)
 		}
@@ -194,12 +195,12 @@ func TestStoreOptionsUnmarshaling(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			opts := &factory.NewStoreOptions{
+			opts := &store.NewOptions{
 				Type:       registryStoreType,
 				Parameters: test.params,
 			}
 
-			store, err := factory.NewStore(opts)
+			store, err := store.New([]*store.NewOptions{opts}, nil)
 			if (err != nil) != test.expectErr {
 				t.Errorf("expected error: %v, got: %v", test.expectErr, err)
 			}
@@ -269,14 +270,14 @@ func TestCredentialProviderIntegration(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			opts := &factory.NewStoreOptions{
+			opts := &store.NewOptions{
 				Type: registryStoreType,
 				Parameters: map[string]interface{}{
 					"credential": test.credConfig,
 				},
 			}
 
-			store, err := factory.NewStore(opts)
+			store, err := store.New([]*store.NewOptions{opts}, nil)
 			if (err != nil) != test.expectErr {
 				t.Errorf("expected error: %v, got error: %v", test.expectErr, err)
 			}
