+++ internal/verifier/factory_test.go
@@ -13,20 +13,25 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package factory
+package verifier
 
 import (
+	"context"
 	"testing"
 
 	"github.com/notaryproject/ratify-go"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+	"github.com/stretchr/testify/assert"
 )
 
 const (
 	testType = "test-type"
 	testName = "test-name"
+	mockName = "mock-name"
+	mockType = "mock-type"
 )
 
-func createVerifier(_ *NewVerifierOptions, _ []string) (ratify.Verifier, error) {
+func createVerifier(_ *NewOptions, _ []string) (ratify.Verifier, error) {
 	return nil, nil
 }
 func TestRegisterVerifierFactory(t *testing.T) {
@@ -72,19 +77,19 @@ func TestRegisterVerifierFactory(t *testing.T) {
 
 func TestNewVerifier(t *testing.T) {
 	t.Run("Creating a verifier with empty name or type", func(t *testing.T) {
-		_, err := NewVerifier(&NewVerifierOptions{Name: "", Type: testType}, nil)
+		_, err := New(&NewOptions{Name: "", Type: testType}, nil)
 		if err == nil {
 			t.Errorf("Expected error when creating a verifier with empty name, but got none")
 		}
 
-		_, err = NewVerifier(&NewVerifierOptions{Name: testName, Type: ""}, nil)
+		_, err = New(&NewOptions{Name: testName, Type: ""}, nil)
 		if err == nil {
 			t.Errorf("Expected error when creating a verifier with empty type, but got none")
 		}
 	})
 
 	t.Run("Creating a verifier with unregistered type", func(t *testing.T) {
-		_, err := NewVerifier(&NewVerifierOptions{Name: testName, Type: "unregistered-type"}, nil)
+		_, err := New(&NewOptions{Name: testName, Type: "unregistered-type"}, nil)
 		if err == nil {
 			t.Errorf("Expected error when creating a verifier with unregistered type, but got none")
 		}
@@ -96,10 +101,85 @@ func TestNewVerifier(t *testing.T) {
 			delete(registeredVerifiers, testType)
 		}()
 
-		opts := &NewVerifierOptions{Name: testName, Type: testType}
-		_, err := NewVerifier(opts, nil)
+		opts := &NewOptions{Name: testName, Type: testType}
+		_, err := New(opts, nil)
 		if err != nil {
 			t.Errorf("Did not expect error when creating a verifier with registered type, but got: %v", err)
 		}
 	})
+}
+
+type mockVerifier struct{}
+
+func (m *mockVerifier) Name() string {
+	return mockName
+}
+func (m *mockVerifier) Type() string {
+	return mockType
+}
+func (m *mockVerifier) Verifiable(_ ocispec.Descriptor) bool {
+	return true
+}
+
+func (m *mockVerifier) Verify(_ context.Context, _ *ratify.VerifyOptions) (*ratify.VerificationResult, error) {
+	return &ratify.VerificationResult{}, nil
+}
+
+func createMockVerifier(_ *NewOptions, _ []string) (ratify.Verifier, error) {
+	return &mockVerifier{}, nil
+}
+
+func TestNewVerifiers(t *testing.T) {
+	RegisterVerifierFactory("mock-type", createMockVerifier)
+	tests := []struct {
+		name          string
+		opts          []*NewOptions
+		expectErr     bool
+		expectedCount int
+	}{
+		{
+			name:          "no options provided",
+			opts:          []*NewOptions{},
+			expectErr:     true,
+			expectedCount: 0,
+		},
+		{
+			name: "error during NewVerifier",
+			opts: []*NewOptions{
+				{
+					Name:       "notation-1",
+					Type:       "notation",
+					Parameters: map[string]interface{}{},
+				},
+			},
+			expectErr: true,
+		},
+		{
+			name: "single valid option",
+			opts: []*NewOptions{
+				{
+					Name: mockName,
+					Type: mockType,
+				},
+			},
+			expectErr:     false,
+			expectedCount: 1,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			verifiers, err := NewVerifiers(tt.opts, nil)
+			if tt.expectErr {
+				assert.Error(t, err)
+				assert.Nil(t, verifiers)
+			} else {
+				assert.NoError(t, err)
+				assert.Len(t, verifiers, tt.expectedCount)
+				for _, verifier := range verifiers {
+					assert.Implements(t, (*ratify.Verifier)(nil), verifier)
+				}
+			}
+		})
+	}
 }
