+++ internal/verifier/notation/register_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 
 	"github.com/notaryproject/notation-go/verifier/truststore"
-	"github.com/notaryproject/ratify/v2/internal/verifier/factory"
+	"github.com/notaryproject/ratify/v2/internal/verifier"
 	"github.com/notaryproject/ratify/v2/internal/verifier/keyprovider"
 )
 
@@ -66,12 +66,12 @@ func TestNewVerifier(t *testing.T) {
 
 	tests := []struct {
 		name      string
-		opts      *factory.NewVerifierOptions
+		opts      *verifier.NewOptions
 		expectErr bool
 	}{
 		{
 			name: "Unsupported params",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type:       verifierTypeNotation,
 				Name:       testName,
 				Parameters: make(chan int),
@@ -80,7 +80,7 @@ func TestNewVerifier(t *testing.T) {
 		},
 		{
 			name: "Malformed params",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type:       verifierTypeNotation,
 				Name:       testName,
 				Parameters: "{",
@@ -89,7 +89,7 @@ func TestNewVerifier(t *testing.T) {
 		},
 		{
 			name: "Missing trust store options",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type:       verifierTypeNotation,
 				Name:       testName,
 				Parameters: options{},
@@ -98,7 +98,7 @@ func TestNewVerifier(t *testing.T) {
 		},
 		{
 			name: "Invalid trust store type",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type: verifierTypeNotation,
 				Name: testName,
 				Parameters: options{
@@ -113,7 +113,7 @@ func TestNewVerifier(t *testing.T) {
 		},
 		{
 			name: "Duplicate trust store type",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type: verifierTypeNotation,
 				Name: testName,
 				Parameters: options{
@@ -133,7 +133,7 @@ func TestNewVerifier(t *testing.T) {
 		},
 		{
 			name: "Non-registered key provider",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type: verifierTypeNotation,
 				Name: testName,
 				Parameters: options{
@@ -149,7 +149,7 @@ func TestNewVerifier(t *testing.T) {
 		},
 		{
 			name: "Key provider that would fail on GetCertificates (lazy loading)",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type: verifierTypeNotation,
 				Name: testName,
 				Parameters: options{
@@ -167,7 +167,7 @@ func TestNewVerifier(t *testing.T) {
 		},
 		{
 			name: "Valid notation options",
-			opts: &factory.NewVerifierOptions{
+			opts: &verifier.NewOptions{
 				Type: verifierTypeNotation,
 				Name: testName,
 				Parameters: options{
@@ -185,7 +185,7 @@ func TestNewVerifier(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			_, err := factory.NewVerifier(test.opts, nil)
+			_, err := verifier.New(test.opts, nil)
 			if test.expectErr != (err != nil) {
 				t.Fatalf("Expected error: %v, got: %v", test.expectErr, err)
 			}
